Global config fixes #60. Carefully detach iput

+++ taca/storage/storage.py
@@ -6,9 +6,10 @@ import shutil
 import time
 
 from taca.log import LOG
-from taca.utils import filesystem
+from taca.utils.config import CONFIG as config
+from taca.utils import filesystem, misc
 
-def cleanup(config, days):
+def cleanup(days):
     for data_dir in config.get('storage').get('data_dirs'):
         with filesystem.chdir(data_dir):
             for run in [r for r in os.listdir(data_dir) if re.match(filesystem.RUN_RE, r)]:
@@ -22,7 +23,7 @@ def cleanup(config, days):
                         LOG.info('RTAComplete.txt file exists but is not older than {} day(s), skipping run {}'.format(str(days), run))
 
 
-def archive_to_swestore(config, days, run=None):
+def archive_to_swestore(days, run=None):
     # If the run is specified in the command line, check that exists and archive
     if run:
         run = os.path.basename(run)
@@ -69,14 +70,14 @@ def _archive_run(config, run):
         :param str dest: Destination directory in Swestore
         :param bool remove: If True, remove original file from source
         """
-        if not filesystem.is_in_swestore(f):
+        if not filesystem.is_in_swestore(f) and not misc.exists_process_with_text(run):
             LOG.info("Sending {} to swestore".format(f))
-            misc.call_external_command('iput -K -P {file} {dest}'.format(file=f, dest=dest),
+            misc.call_external_command_detached('iput -K -P {file} {dest}'.format(file=f, dest=dest),
                     with_log_files=True)
             LOG.info('Run {} sent correctly and checksum was okay.'.format(f))
         else:
-            LOG.warn('Run {} is already in Swestore, not sending it again'.format(f))
-        if remove:
+            LOG.warn('Run {} is already in Swestore or currently being archived, not sending it again'.format(f))
+        if remove and not misc.exists_process_with_text(run):
             LOG.info('Removing run'.format(f))
             os.remove(f)
